+++ src/main/java/Duch.java
@@ -1,5 +1,10 @@
 import java.util.Scanner;
 
+import Duch.Parser;
+import Duch.Storage;
+import Duch.TaskList;
+import Duch.Ui;
+
