+++ src/main/java/Duch/DuchException.java
@@ -1,0 +1,7 @@
+package Duch;
+
+public class DuchException extends Exception {
+    DuchException(String message) {
+        super(message);
+    }
+}
