+++ src/main/java/Duch/Storage.java
@@ -1,10 +1,13 @@
+package Duch;
+
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardOpenOption;
 import java.util.Scanner;
-import Task.*;
+
+import Duch.Task.*;
