+++ src/main/java/Duch/Task/Date.java
@@ -1,4 +1,4 @@
-package Task;
+package Duch.Task;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
