--- src/main/java/DuchException.java
@@ -1,5 +1,0 @@
-public class DuchException extends Exception {
-    DuchException(String message) {
-        super(message);
-    }
-}
