--- src/main/java/Task/Task.java
@@ -1,24 +1,0 @@
-package Task;
-
-public abstract class Task {
-    String task;
-    boolean isDone;
-
-    public Task(String task, boolean done) {
-        this.task = task;
-        isDone = done;
-    }
-
-    public Task(String task) {
-        this.task = task;
-        isDone = false;
-    }
-
-    public void setDone(boolean val) {
-        isDone = val;
-    }
-
-    public abstract String toString();
-
-    public abstract String toStore();
-}
