+++ text-ui-test/EXPECTED.TXT
@@ -25,4 +25,16 @@ Please input your dates in d/MM/yyyy HHmm
     3.[E][ ] project meeting (from: Aug-20-2024 1700 to: Sep-20-2024 1700)
     ---------------------------------------------
 
+    ---------------------------------------------
+    /by date is not found
+    ---------------------------------------------
+
+    ---------------------------------------------
+    The description of todo cannot be empty
+    ---------------------------------------------
+
+    ---------------------------------------------
+    /from date (and/or) /to date is not found
+    ---------------------------------------------
+
 Bye. Hope to see you again
