+++ text-ui-test/input.txt
@@ -2,4 +2,7 @@ todo borrow book
 deadline return book /by 20/08/2024 1700
 event project meeting /from 20/08/2024 1700 /to 20/09/2024 1700
 list
+deadline return money
+todo
+event holiday /fro 20/08/2024 1700
 bye
